requirements.txt
attrs==20.3.0
Automat==20.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
constantly==15.1.0
cryptography==3.4.4
cssselect==1.1.0
Flask==1.1.1
future==0.18.2
hyperlink==21.0.0
idna==2.10
incremental==17.5.0
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
lxml==4.6.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
numpy==1.20.1
opencv-python==4.5.1.48
parsel==1.6.0
pdfkit==0.6.1
Pillow==8.1.2
Protego==0.1.16
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
python-telegram-bot==12.2.0
queuelib==1.5.0
requests==2.25.1
Scrapy==2.4.1
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
tornado==6.1
Twisted==20.3.0
urllib3==1.26.3
w3lib==1.22.0
Werkzeug==1.0.1
zope.interface==5.2.0
